add getCommandType and deleteCommandType controllers

+++ models/CommandType.ts
@@ -24,10 +24,12 @@ export interface ICommandType extends Document{
 }
 
 CommandTypeSchema.post('findOneAndDelete', async (doc: ICommandType) => {
-	try {
-		await Command.deleteMany({CommandType: doc._id})
-	} catch (e) {
-		console.log(e);
+	if (doc) {
+		try {
+			await Command.deleteMany({CommandType: doc._id})
+		} catch (e) {
+			console.log(e);
+		}
 	}
 })
